+++ app/repository/_component/type.ts
@@ -1,0 +1,10 @@
+export type RegistrationProps = {
+    index: number;
+    stepDesc: string;
+    imageSrc: string;
+    imageHeight: number;
+    secondImageSrc?: string;
+    secondImageHeight?: number;
+    imageDesc?: string;
+    secondImageDesc?: string;
+};
